==> Scripts/1 GTFS Arcpy Shapes to Features.ipynb <==
## This top cell is the correct one

#GTFSShapesToFeatures_conversion Test
import arcpy
shapes = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/GTFS/oct1/shapes.txt'
outclass = 'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Testing.gdb/GTFSShapesToFeatures'
arcpy.conversion.GTFSShapesToFeatures(shapes, outclass)

#This seems to create a polyline set of existing shapes simply from a shapes.txt file

## Tested, not ideal

#GenerateShapesFeaturesFromGTFS_conversion Test
GTFSFolder = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/GTFS/oct1NoShapes'
outshapes = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Testing.gdb/outshapes'
outstops = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Testing.gdb/outstops'
outtrips =r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/outtrips.txt'
arcpy.conversion.GenerateShapesFeaturesFromGTFS(GTFSFolder, outshapes, outstops, outtrips,[])


#You seem to be able to generate straight lines OR along a network with this tool 
#(although without the network there are some straight and bent). 
#Also creates transit stop layout

#FeaturesToGTFSShapes Test
inshapes = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Testing.gdb/outshapes'
instops = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Testing.gdb/outstops'
intrips =r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/outtrips.txt'
# above 3 are outputs of the previous tool, GenerateShapesFeaturesFromGTFS

OriginalStopTimes = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/GTFS/oct1/stop_times.txt'
#this is the original stop times file from the base GTFS set


NewShapes = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/shapes_new.txt'
NewStopTimes = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/stop_times_new.txt'

arcpy.conversion.FeaturesToGTFSShapes(inshapes, instops, intrips, OriginalStopTimes, NewShapes, NewStopTimes, 'METERS')
