==> Scripts/2 Comparing GTFS Files.ipynb <==
#file names in GTFS
fileslist = ['agency','calendar','calendar_dates','routes','shapes','stop_times','stops', 'trips']

#for each file name compare the 5th and 6th (or any two adjacent days)
for each in fileslist: 
    first_file = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/GTFS/oct5/' + each +'.txt'
    second_file = r'C:/Users/Cole/Desktop/Spring2022/AVLResearch/Data/GTFS/oct6/' + each +'.txt'
    file1 = open(first_file, "r")
    file2 = open(second_file, "r")
    print(each)
    i = 0
    #row by row
    for l1 in file1:
        i += 1
        for l2 in file2:
            if l1 == l2:
                pass
                #"Line is same ", i,": ")
            else:
                print("Line ", i,": ")
                #show lines that are different
                print("File 1: ", l1)
                print("File 2: ", l2)
            break
    #close files
    file1.close()
    file2.close()
