==> Scripts/3 GTFS Joining.ipynb <==
### Imports

import pandas as pd
import os
#import geopandas as gp

# Set the workspace environment to master folder
x=os.chdir('C:\\Users\\Cole\\Desktop\\Spring2022\\AVLResearch')

### Define the GTFS inputs

#open all files into dataframes
trips_txt = pd.read_csv("Data/GTFSoct1/trips.txt", index_col = False)
shapes_txt = pd.read_csv("Data/GTFSoct1/shapes.txt", index_col = False)
stop_times_txt= pd.read_csv("Data/GTFSoct1/stop_times.txt", index_col = False)
stops_txt= pd.read_csv("Data/GTFSoct1/stops.txt", index_col = False)
calendar_txt= pd.read_csv("Data/GTFSoct1/calendar.txt", index_col = False)
calendar_dates_txt= pd.read_csv("Data/GTFSoct1/calendar_dates.txt", index_col = False)

### Query route (calendar query)

#find the route ID for the route name you want
routes_txt = pd.read_csv("Data/GTFSoct1/routes.txt")
route_name = 'A Line'
route_row = routes_txt.loc[routes_txt['route_short_name'] == route_name]
route_row.to_csv('Data/GTFS_Query/routes_select.txt')
#find the service_ids for calendar rows you want
cal_select = calendar_txt.loc[calendar_txt['service_id'].isin(['AUG16-MVS-BUS-Weekday-01',
                                                               'AUG16-MVS-BUS-Saturday-01',
                                                               'AUG16-MVS-BUS-Sunday-01'])] 
cal_select.to_csv('Data/GTFS_Query/cal_select.txt', index=False )

### Carry query through 

#select trips using calendar_select and route_select frames
trips_select_pre = trips_txt[trips_txt["route_id"] == "921-95"]
trips_select = pd.merge(cal_select, trips_select_pre, how = 'left', on = 'service_id')
trips_select = trips_select[['route_id',
                             'service_id',
                             'trip_id',
                             'trip_headsign',
                             'direction_id',
                             'shape_id']]
trips_select.to_csv('Data/GTFS_Query/trips_select.txt', header=1, index=False  )

#select shape to trip on shape_id
shapes_select = pd.merge(trips_select,shapes_txt, how = "left", on = "shape_id")
shapes_select = shapes_select[['route_id',
                              'trip_id',
                              'service_id',
                              'trip_headsign',
                              'shape_id',
                              'shape_pt_lat',
                              'shape_pt_lon',
                              'shape_pt_sequence']]
shapes_select.to_csv('Data/GTFS_Query/shapes_select.txt', header=1, index=False  )

#select stoptimes to trip on trip_id
stop_times_select = pd.merge(trips_select,stop_times_txt, how = "left", on = "trip_id")
stop_times_select = stop_times_select[['route_id',
                                       'trip_id',
                                       'service_id',
                                       'trip_headsign',
                                       'shape_id',
                                       'arrival_time',
                                       'departure_time',
                                       'stop_id',
                                       'stop_sequence']]
#stop_times_select = stop_times_select.drop(stop_times_select.columns[:8], axis=1)
#stop_times_select
stop_times_select.to_csv('Data/GTFS_Query/stop_times_select.txt', header=1, index=False )

#select just stop locations by stop_id
stops_select = pd.merge(stop_times_select,stops_txt, how = "left", on = "stop_id")
stops_select = stops_select.drop_duplicates(subset='stop_id', keep="first")
stops_select = stops_select[['stop_id',
                             'stop_name',
                             'stop_desc',
                             'stop_lat',
                             'stop_lon',
                             'stop_sequence']] 
stops_select
stops_select.to_csv('Data/GTFS_Query/stops_select.txt', header=1, index=False  )
